Alerts: Turn heading into BEM element
Removes the redefinition of GovUK styles as that is incorrect and will
affect the rest of the website. Instead it defines a BEM element and
styles that directly.

I think there's also a question of whether this should really be an
`<h2>` of whether it should just be a `<div>`.