Add Timeline component (#66)
Initial build of Timeline component using React. The timeline can be created with each event as a separate item within an array of items. Each item consists of an event name, user name (optional), date and time, and description.
In the HMCTS Timeline component, there is the ability to add documents to the timeline however we feel that this is not beneficial to us at this time.

* further changes to Timeline component. Tests now working and output shown on Storybook

* Changes to responsive margins and text formatting to bring it in line with govuk styling, following comments against my PR.

* Fiex to pnpm-lock error when committing changes.

* after running tests

* Changes to "standard" story to enable UI test to run

* Have finalised unit tests for timeline component. Have also addressed other issues raised in PR.

* Changes to styling to remove nested media queries

* Add test to count child elements. Updated documentation for timeline component.

* Remove background colour on mouseover

* Minor changes to padding and margins

* Remove unnecessary test and story for if no items prop provided

Co-authored-by: haggishunt56 <whatevermaanhotmail.com>
Co-authored-by: Daniel A.C. Martin <daniel-ac-martin@users.noreply.github.com>