refactor: actions as objects
Expect an array of objects, specifying a `text` and `href` property, to
make the component slightly nicer to consume.
Also don't force the ` - ` separator between the status and extra
contents, just use a single space (so that the contents don't run into
each other) to allow the calling context to decide what separator (if
any) should be used.