eGates and ePassport
New entries added:

Added under 'Biometric passport' and 'ePassport':
Use ‘passport’ if the type of passport does not matter. When referring to passports with a chip, use ‘biometric passport’ instead of ‘ePassport’.

eGates
When using ‘eGates’ with the public, explain what it means the first time you use it. For example, ‘you will have the option of using an eGate (electronic passport gate)’.