CI: Handle package lock-file in CI
Tracks the package lock-file in a special location and .gitignore's the
standard location. This allows us to avoid having the lock-file being
included in pull requests and the extra problems that can create.

Handing the lock-file in CI should also be more secure and it helps
prevent bad things from being injected.